+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Block metadata page: added "Show as YAML" toggle
 ### Fixed
 - Data tab: fixed view for "Saved queries" block
 - Settings tab: active tab is highlighted correctly